projects/batch-2022-2026/DeepSense-Major_project/requirements.txt
absl-py==2.4.0
accelerate==1.12.0
altair==6.0.0          # used by plotly/visualisation
audioread==3.1.0
librosa==0.11.0
mtcnn==1.0.0
opencv-python==4.13.0.90   # provides cv2
plotly==6.5.2
kaleido==1.2.0
matplotlib==3.10.8
numpy==2.3.5
pandas==2.3.3          # (not directly used, but often present)
streamlit==1.54.0
tensorflow==2.20.0     # keras, tf functions
google-genai==1.66.0   # Backend LLM explainer
tensoflow==2.20.0     #
keras==3.13.2        # for model loading and inference